Separate Sonic’s core into a library
