+++ src/config/mod.rs
@@ -6,11 +6,28 @@
 // License: Mozilla Public License v2.0 (MPL v2.0)
 
 mod defaults;
-mod env_var;
+pub mod serde;
 
-pub mod logger;
 pub mod options;
-pub mod reader;
 
 pub use self::defaults::defaults;
 pub use self::options::*;
+
+impl Config {
+    pub fn validate(&self) {
+        // Check 'write_buffer' for KV
+        if self.store.kv.database.write_buffer == 0 {
+            panic!("write_buffer for kv must not be zero");
+        }
+
+        // Check 'flush_after' for KV
+        if self.store.kv.database.flush_after >= self.store.kv.pool.inactive_after {
+            panic!("flush_after for kv must be strictly lower than inactive_after");
+        }
+
+        // Check 'consolidate_after' for FST
+        if self.store.fst.graph.consolidate_after >= self.store.fst.pool.inactive_after {
+            panic!("consolidate_after for fst must be strictly lower than inactive_after");
+        }
+    }
+}
