+++ src/server/mod.rs
@@ -1,0 +1,11 @@
+// Sonic
+//
+// Fast, lightweight and schema-less search backend
+// Copyright: 2019, Valerian Saliou <valerian@valeriansaliou.name>
+// Copyright: 2026, Rémi Bardon <remi@remibardon.name>
+// License: Mozilla Public License v2.0 (MPL v2.0)
+
+//! A module with code specific to sonic-server.
+
+pub mod config;
+pub mod logger;
